Modifying genData to combine evc and mt
