+++ brain/computeRate_mt.m
@@ -1,4 +1,4 @@
-function firing_rate = computeRate(resp,stim,cdir,data,settings)
+function firing_rate = computeRate_mt(resp,stim,cdir,data,settings)
 
 % use the direction of each cell as a multiplier on this stimulus
 fireScale = normpdf(abs(data(:,4)-cdir),0,settings.mt.dirSigma)/settings.mt.maxFire;
@@ -26,7 +26,7 @@ firing_rate(firing_rate<0) = 0;
 % firing_rate = firing_rate * (settings.def_fire+settings.max_fire) / max(firing_rate(:));
 
 % test (if needed)
-figure; imagesc(squeeze(reshape(firing_rate(:,2001),51,51))); colormap('gray'); colorbar
+% figure; imagesc(squeeze(reshape(firing_rate(:,2001),51,51))); colormap('gray'); colorbar
 
 % collapse
 firing_rate = (round(firing_rate(:)*10)/10)';
