+++ brain/data/data.js
@@ -1,1 +1,1 @@
-module.exports = {  A: require('./data1.json'),  B: require('./data2.json'),  C: require('./data3.json'),  D: require('./data4.json'),  E: require('./data5.json'),  F: require('./data6.json'),  G: require('./data7.json'),  H: require('./data8.json'),  I: require('./data9.json'),  J: require('./data10.json'),  K: require('./data11.json'),  L: require('./data12.json'),  M: require('./data13.json'),  N: require('./data14.json'),  O: require('./data15.json'),  P: require('./data16.json'),  Q: require('./data17.json'),  R: require('./data18.json'),}
+module.exports = {  A: require('./data_mt1.json'),  B: require('./data_mt2.json'),  C: require('./data_mt3.json'),  D: require('./data_mt4.json'),}
