+++ brain/isfile.m
@@ -1,0 +1,2 @@
+function is = isfile(file)
+is = exist(file,'file');
